chore: move drink detail content to another file
